requirements.txt
appdirs==1.4.3
decorator==4.0.11
dj-database-url==0.4.2
Django==1.10.6
factory-boy==2.8.1
Faker==0.7.10
gunicorn==19.7.1
ipdb==0.10.2
ipython==5.3.0
ipython-genutils==0.2.0
Markdown==2.6.8
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
whitenoise==3.3.0
